pyproject.toml
[project]
name = "hakura-pre-guidelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "guardrails-ai==0.6.0",
    "guardrails-api-client==0.4.0a1",
    "openai>=1.57.0",
    "streamlit>=1.40.2",
]
